create json encoder module
